Bump manylinux image to 2014
Since upstream the PYPA has also dropped support for manylinux2010, and
we need to use the manylinux2014 image for Python 3.11. When combined
with the next version of Rust raising the GLIBC requirement we're at the
point where we can't realistically support manylinux2010 anymore. This
commit updates the documentation, release notes, and the build scripts
to use manylinux2014 for all the linux wheels we publish.