Add support for Python 3.11 (#683)
* Add support for Python 3.11

For the upcoming 0.12.0 release it would be good to add support for
python 3.11. While this hasn't been released yet it's currently a
release candidate with a frozen C API. We're likely to release rustworkx
0.12.0 prior to the Python 3.11 release. In order to avoid having to
have rustworkx be a blocker for any downstream packages looking to adopt
3.11 this commit proactively enables support for testing and building
with Python 3.11 off the latest rc in order to ensure we're building
binaries compatible with 3.11 for the 0.12.0 release.

* Explicitly use rc2 in ci config

* DNM: only run wheel jobs in CI to test new config

* Pin Rust to 1.63 on linux wheel jobs

In the upcoming Rust 1.64.0 release the minimum glibc version is being
raised to 2.17 which is higher than the minimum supported version in
manylinux 2010 which is our current Python wheel target environment. To
continue supporting manylinux 2010 for Rustworkx 0.12.0 we need to pin
the rust version to avoid implicitly increasing the minimum after rust
1.64 is released. This also adds a release note documenting that this is
the last release that supports manylinux 2010.

* Update test command

* Bump manylinux image to 2014

Since upstream the PYPA has also dropped support for manylinux2010, and
we need to use the manylinux2014 image for Python 3.11. When combined
with the next version of Rust raising the GLIBC requirement we're at the
point where we can't realistically support manylinux2010 anymore. This
commit updates the documentation, release notes, and the build scripts
to use manylinux2014 for all the linux wheels we publish.

* Revert rust compiler pin

Since we've moved to manylinux2014 for all linux builds we no longer
need to pin the rust compiler to avoid requiring a newer glibc. This
commit removes the version pin when installing rust in the cibuildwheel
environment.

* Revert "DNM: only run wheel jobs in CI to test new config"

This reverts commit 2d26af513391a39f5c66a13cb3cfb3956163847a.