Add index_output option to layers() function (#684)
* Add index_output option to layers() function

This commit adds a new keyword argument, index_output, to the layers()
function which is used to return a list of list of node indices
representing the layers in the DAG instead of a list of list of node
data payloads. The default output does not change, but for users that
prefer to work with indices (which is probably a more natural return
format for the function) this option allows them to work with indices.

Closes #646

* Add missing doc indent