Add Johnson's algorithm for cycles in directed graph (#633)
* Add Johnson's algorithm for cycles in directed graph

This commit adds a new function, simple_cycles(),  which is an
implementation of Johnson's algorithm [1] for finding all the simple
cycles in a directed graph. The implementation is based on the
non-recursive implementation from NetworkX. [2]

[1] https://doi.org/10.1137/0204007
[2] https://github.com/networkx/networkx/blob/networkx-2.8.4/networkx/algorithms/cycles.py#L98-L222

Closes: #622

* Remove test debugging

* Fix lint

* Add more tests

* Apply suggestions from code review

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Fix rustfmt

* Make return from simple_cycles a iterator

This commit converts the simple_cycles function to a Python class that
implements the iterator protocol. Each step of the the iterator will
compute the next cycle using Johnson's algorithm and return a
NodeIndices object representing the cycle.

* Use IndexSet::pop() instead of set_pop()

* Fix iterator behavior

* Fix MSRV support

* Use mem::take() instead of .drain(..).collect()

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>
Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>