Fix retworkx compat artifact CI job
Before pushing the 0.12.0 tag I noticed the CI configuration for the
retworkx release artifacts was broken and wouldn't run because of a
dependency on the rustworkx jobs that was defined in a separate
workflow. This commit updates the CI config by adding the retworkx
artifact job to the wheels workflow so we can express the dependency on
rustworkx being published first.