Add cargo publish job for rustworkx-core (#694)
This commit adds a CI job to automatically publish rustworkx-core to
crates.io on releases. In the past I have just manually run cargo
publish locally but this is error prone and depends on me doing the
correct thing. Having it run automatically in CI limits the chances of
something going wrong.