Add build_transitive_closure_dag() to rustworkx-core
This commit adds a new function to rustworkx-core for building a
transitive closure inplace out of an input DAG. The function takes
ownership of the input graph and will mutate it to add edtges to make it
a transitive closure before returning it. This is then used internally
by the retworkx python function transitive_clsoure_dag().