Bump MSRV to 1.56.1 and update dependencies and Rust Edition
The previous minimum support Rust version of 1.48 was originally
selected to match the latest version of Rust available through Debian
stable. This was done primarily to support the piwheels build
environemnt which installs rustc via Debian packages. However, the 1.48
MSRV is causing a lot development friction as it's quite old and is
holding back a lot of dependencies. When weighing the relatively tiny
piwheels user base (for an unsupported platform) against this it's
probably time to decouple our MSRV for rustworkx from that and pick a
version that makes the most sense for our development needs.

This commit raises the minimum supported rust version to 1.56.1 and then
bumps all our dependencies held back because of the previous MSRV. At
the same time the rust edition used for rustworkx is changed to be
edition 2021 instead of edition 2018. This will enable us to leverage
newer rust syntax now.

Fixes #716