Add longest_simple_path() function to rustworkx
This commit adds a new function longest_simple_path() which is used to
find the longest simple path in the graph. It internally is running
basically the same algorithm as all_pairs_all_simple_paths but instead
of building a giant nested mapping of all the simple paths found in the
graph it instead filters the return to just the longest path found. This
is useful if you're quickly trying to find the longest path in the graph
as it avoids the overhead of building the return object and then
iterating over it again.