Fix support of sequence protocol for returned lists (#730)
* Fix support of sequence protocol for returned lists

Add the `sequence` option to PyO3's `pyclass`, so that the `sq_length`
slot is implemented [1]. Implementing this method is required for
sequence types, or Python C API functions like `PySequence_Size` will
fail with an error.

The `reversed` built-in method relies on `PySequence_*` methods. A test
reversing `NodeIndices` is added to guard against future violations of
the sequence protocol.

Fixes #696.

[1]: https://github.com/PyO3/pyo3/pull/2567

* Add release note

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>