Fix compatibility with priority-queue 1.3.0 (#745)
In priority-queue 1.3.0 compilation of rustworkx-core's min_cut module
will fail because of the lack of a semicolon on the change_priority_by()
call. This wasn't an issue with priority-queue 1.2.0 and for rustworkx
development we are using priority-queue 1.2.0 because the priority queue
1.3.0 requires a rust version newer than our msrv. However, adding a
semicolon to fix this compilation issue is also still compatibile with
priority-queue 1.2.0. This commit does that to fix compatibility with
newer versions of rustworkx-core that want to leverage the latest
version of priority queue where they don't have the same MSRV
constraints as rustworkx.

Fixes #744