Deduplicate wheel job definitions and update cibuildwheel
This commit simplifies the wheel action configuration and deduplicates
much of the job definitions. This is done primarily by unifying the
cibuildwheel configuration in the pyproject.toml to set default values
for all the common options. Then the separate duplicated install and run
job stages are switched over to the native action published by
cibuildwheel which deduplicates the configuration further. At the same
time this commit bumps the configuration to use the latest version of
cibuildwheel.