Only publish wheel for retworkx compat package (#749)
This commit fixes the retworkx compat package publish job. The first
issue is that the sdist for the retworkx package is not valid, we
trigger the creation of the compat package by an env variable when
calling setup.py/setuptools/builder/etc which triggers change the
package name and other attributes. However if we create an sdist while
setting the env variable this just bundles the setup.py and other source
files in a tarball and when the sdist is installed rustworkx would be
installed. So for the compat package we should only publish a pure
python wheel that will have the actual compat shim. The other small
change is switch the package to use python 3.10 instead of 3.7, which
goes EoL in the near future.

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
(cherry picked from commit 100d0dc87707e53e9510b6ff13045ad802d0d14f)