Move cycle_graph generator to rustworkx-core (#758)
* Move cycle_graph generator to rustworkx-core

This commit migrates the cycle-graph generator to rustworkx-core. It
makes the function generic so that it can be used with any petgraph
graph that's buildable and node indexable. This is primarily to serve as
a pattern for migrating all the generator functions we currently have to
rustworkx-core. This pattern should be usable for any generator function
we have.

Part of #751

* Apply suggestions from code review

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>

* Run cargo fmt

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>