Get list of supported image formats from `dot` instead of hardcoding
This PR replaces a list of image formats supported by graphviz that is hardcoded in
rustworkx/visualization/graphviz.py with a list that is taken dynamically from the
executable `dot`. With the version of `dot` that I tested, this list had 50 entries,
while the hardcoded list had 35.

The immediate reason for this PR is due to a failure in the test suite despite having
a working `dot` on the PATH. Apparently `dot` can be compiled with varying support
for image formats. In addition on some linux distributions, including some Ubuntu and
some Fedora, the package `graphviz` does not support jpeg. But, installing the
additional package `graphviz-devel` does add jpeg support. Before this PR, this would
cause an untrapped error and a failure in the test suite, because only the presence
of `dot` is checked, not which images it supports. With this PR, if your `dot` does
not support `jpeg` then the test for writing a `jpeg` will be skipped.

This PR also implements an error message that if both pillow and graphviz are missing
says both are missing. Before, an error would be thrown with one message, then after
installing the dependency, an error would be thrown with the second.