Improve documentation for graph_greedy_color
This commit improves the documentation for the graph_greedy_color
function. Previously, the details on the function and the algorithm it
implemented where a bit sparse. This commit expands it by explaining the
source for the algorithm, making it clear it's not always going to
return an optimal solution, and also adding an example.