Pin qiskit_sphinx_theme version
The recently released 1.11.x release series of the sphinx theme is
incompatible with the customization we made to the templates for the
theme to remove the top bar. The recommendation from the theme
package maintainers was to just pin the theme version to the old version
if the modifications are required. This should unblock the CI docs failure
we've seen for the past couple of weeks which was caused by an
incompatibility with our template files.

Longer term, I think we might want to migrate to a different theme
package for the rustworkx docs. The direciton of the qiskit docs theme
is to be more tightly coupled with the evolving Qiskit ecosystem which
makes a lot of sense for the other consumers of the theme package. But for
rustworkx we typically have tried to keep it independent as it's a base
dependency for Qiskit, so selecting a different theme is probably the
best path forward to avoid tension like this in the future.