Improve documentation for graph_greedy_color (#857)
* Improve documentation for graph_greedy_color

This commit improves the documentation for the graph_greedy_color
function. Previously, the details on the function and the algorithm it
implemented where a bit sparse. This commit expands it by explaining the
source for the algorithm, making it clear it's not always going to
return an optimal solution, and also adding an example.

* Update src/coloring.rs

* Change to use note directive

* Update src/coloring.rs

Co-authored-by: Julien Gacon <gaconju@gmail.com>

---------

Co-authored-by: Julien Gacon <gaconju@gmail.com>