Add longest_simple_path() function to rustworkx (#731)
* Add longest_simple_path() function to rustworkx

This commit adds a new function longest_simple_path() which is used to
find the longest simple path in the graph. It internally is running
basically the same algorithm as all_pairs_all_simple_paths but instead
of building a giant nested mapping of all the simple paths found in the
graph it instead filters the return to just the longest path found. This
is useful if you're quickly trying to find the longest path in the graph
as it avoids the overhead of building the return object and then
iterating over it again.

* Fix doc example of the alternative python example

This commit fixes the code example for the python space example of how
to compute the longest path in a graph (at the cose of slower runtime
and a lot more memory).

Co-authored-by: Alexander Ivrii <alexi@il.ibm.com>

* Deduplicate function implementations between graph types

* Simplify logic for selection of longest path

* Further simplify iterator

* Better handle empty path lists

* Remove clone() usage

* Add dedidcated longest path function to rustworkx-core

* Update add-longest-simple-path-afdc4538c49bc38f.yaml

---------

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>
Co-authored-by: Alexander Ivrii <alexi@il.ibm.com>