Bump pyo3 and rust numpy version to 0.19.0 (#887)
* Bump pyo3 and rust numpy version to 0.19.0

Pyo3 0.19.0 and rust-numpy 0.19.0 were just released. This commit
updates the version used in rustworkx to these latest releases. At the
same time this updates usage of text signature for classes that was
deprecated in 0.19.0 release. While not fatal for normal builds this
would have failed clippy in CI because we treat warnings as errors.

* Make _dt optional in iterators

---------

Co-authored-by: Ivan Carvalho <8753214+IvanIsCoding@users.noreply.github.com>