Fix pickle/deepcopy node hole handling
This commit fixes an issue introduced by #589 where in certain cases
node holes in a graph would result in a panic being raised. This was
caused by a logic bug in trying to recreate the holes. Additionally,
there were several places where graph methods removed nodes that the
flag to indicate there were removals would no be set. This commit
fixes all of these issues so that deepcopy/pickle works as expected.