Pin Python version in CI (#905)
* Pin Python version in CI

Recently github updated the cached version of python installed for 3.7
in the CI environment to 3.7.17. This new binary was not built with bz2
support which is causing failures in our CI jobs that run with 3.7.
While we'll be dropping 3.7 support from the main branch (for 0.14.0) in
the near future we still support 3.7 on the stable 0.13.x series. This
commit pins the python version to the previous patch release which was
known to work.

* Use 3.7 on windows

* Fix windows capitalization

* Fix syntax error

(cherry picked from commit 5f73e767fb74ec4ff8e9cf6a2dc6de2c95f30144)