Bump MSRV to 1.64 for 0.14.0
This commit raises the MSRV for rustworkx and rustworkx-core to 1.64.
Previously we had been matching the version of rust included with the
latest debian stable release, which is 1.63 right now. However, some of
our dependencies in rustworkx's latest release has bumped their msrv to
1.64 and instead of waiting for the next debian release it's probably
better to just bump to 1.64. Installing newer versions of rust are
something more users are accustomed to at this point so the need for
debian compatibility isn't as important as it used to be.