Add a section to Python docs about page for rustworkx-core
The main page explaining the rustworkx library in the Python docs
previously didn't make any mention of rustworkx-core. While in general
the Python side is built on top of rustworkx-core and the vast majority
of the users of the Python library don't need to know about the pure
Rust library inside, it is still good to highlight its existence in case
anyone finds the Python library but is really interested in consuming it
from Rust. This commit just adds a small subsection to the about page
explaining the existence of rustworkx-core, its purpose, and where you
can find mroe information about it.