Handle token_swapper impossible swap mapping with Error (#971)
* Handle token_swapper impossible swap mapping with Error

This commit fixes the handling of invalid mapping requests in the
token_swapper rustworkx-core function and the graph_token_swapper
function in rustworkx that uses it. Previously if an invalid mapping was
requested the function would internally panic because it always assumed
there was a path in the graph to fulfill the user requested mapping. However,
because the rustworkx-core function didn't support error returns a
breaking api change is needed to add a result return type.

* Fix formatting

---------

Co-authored-by: Edwin Navarro <enavarro@comcast.net>