Add support for Python 3.12 and musllinux to main
As part of the 0.13.2 release we added support for Python 3.12 and
musllinux to rustworkx. However, these changes did not happen on main
yet. This commit applies the necessary changes to the main branch for
the 0.14.0 and future releases. Also to simplify the configuration of
the cibuildwheel jobs this combines #753 into this PR so that the
configuration is centralized in the pyproject.toml.