Add support for Python 3.12 and musllinux to main (#997)
* Add support for Python 3.12 and musllinux to main

As part of the 0.13.2 release we added support for Python 3.12 and
musllinux to rustworkx. However, these changes did not happen on main
yet. This commit applies the necessary changes to the main branch for
the 0.14.0 and future releases. Also to simplify the configuration of
the cibuildwheel jobs this combines #753 into this PR so that the
configuration is centralized in the pyproject.toml.

* DNM: Test wheel build configuration

* Remove testtools usage from test suite

* Don't require blas on numpy install

* Fix docs typo

* Drop musl on aarch64

* Try installing openblas to fix numpy build

* Fix toml syntax

* Adjust override to use allowed fields only

* Fix toml syntax again

* Switch back to pip command

* Downgrade pp64le to tier 4 and skip tests

* Fix apk command copy paste error

* Raise minimum supported macOS version to 10.12

In Rust 1.74 the Rust programming language is raising their minimum
support macOS version to 10.12, so rustworkx is raising it's supported
version of macOS to match this.

* Tweak test skip regex

* Split arm into 2 jobs

* Revert "DNM: Test wheel build configuration"

This reverts commit 57dd451087e285044d53ccd28194c78215775b57.