Pivot to from flake8 to ruff for Python linting
Since the start of the 0.3.x release of the retworkx project we have
been using flake8 for Python linting and this has worked well for us.
However, recently a new project, ruff [1][2], has shown increasing
popularity becuase it provides the same linting coverage but is
signficantly faster. Also just from a language solidatry PoV ruff is
also written in rust. This commit migrates our linting jobs to use ruff
instead of flake8 and sets up a reasonable set of rules as a starting
point, which includes the rules equivalent to what we were using before
with flake8, pyupgrade, and flake8-pyi, and flake8-quotes.

[1] https://docs.astral.sh/ruff/
[2] https://github.com/astral-sh/ruff