Update pyo3 and numpy to 0.20.0 (#1005)
This commit updates PyO3 to the latest release 0.20.0 [1] and also moves
rust-numpy to 0.20.0 [2] as well. The major change for rustworkx is
official support for Python 3.12.0 final. We were previously using the
3.12 pre-release support in PyO3 to publish rustworkx for 3.12, but
moving to the latest PyO3 officially supports the stable releases of
3.12. The are some small changes needed to adapt to API changes in PyO3,
mostly around updates to the `PyDict.get_item()` api.

[1] https://github.com/PyO3/pyo3/releases/tag/v0.20.0
[2] https://github.com/PyO3/rust-numpy/releases/tag/v0.20.0