Add two_color and is_bipartite (#1002)
* Add two_color and is_bipartite

This commit adds new functions to rustworkx, two_color() and
is_bipartite(), which are used to compute a two coloring for a graph and
then determine if a givn graph is bipartite. The two_color() function is
added to rustworkx-core as the python is_bipartite() function just wraps
it and converts the output to a bool if a two coloring is possible or
not.

This commit is based on top of #998 and will need to be rebased after
that merges.

* Remove special isolates handling

* Expand test coverage