Add stable 0.12 docs to hosted documentation
Since we've migrated our docs to be hosted on github pages we've lost
the historical build of the documentation from the 0.12.1 release. I've
locally built the documentation at 0.12.1 all that was required was
pinning sphinx to 5.3.0 and qiskit-sphinx-theme to 1.9.0 to enable it to
build. This adds this build output to the gh-pages branch for the
stable/0.12 page so we have the historical builds for this version too.