Finalize type signatures and enforce for all future functions
This commit adds the last few missing type hints (primarily the
generators module) and switches the CI configuration over to enforce
that the library is completely type hinted. This makes it a CI error if
a function or method is added that doesn't include type hints. This is
all built on the hard work of @IvanIsCoding who built up all the type
hinting infrastructure and stub files for the majority of rustworkx.
This is just the last piece to enforce the library is fully typed moving
forward.

Co-authored-by: Ivan Carvalho <ivancarv@student.ubc.ca>